Reverted changes commit

+++ xcode-maven-plugin/src/main/java/com/photon/phresco/plugins/xcode/XcodeBuild.java
@@ -253,6 +253,11 @@ public class XcodeBuild extends AbstractMojo implements PluginConstants {
 	 */
 	private String projectType;
 	
+	/**
+	 * @parameter
+	 */
+	private String ciBuild;
+	
 	protected int buildNo;
 	private File buildDirFile;
 	private File buildInfoFile;
@@ -263,6 +268,7 @@ public class XcodeBuild extends AbstractMojo implements PluginConstants {
 	private String dSYMFileName;
 	private String deliverable;
 	private Map<String, Object> sdkOptions;
+	private boolean isCiBuild;
 	
 	/**
 	 * Execute the xcode command line utility.
@@ -275,6 +281,8 @@ public class XcodeBuild extends AbstractMojo implements PluginConstants {
 		getLog().info("basedir xcode " + basedir);
 		getLog().info("baseDir Name Xcode" + baseDir.getName());
 
+		isCiBuild = Boolean.valueOf(ciBuild);
+		
 		try {
 			if(!SdkVerifier.isAvailable(sdk) && !projectType.equals(MAC)) {
 					throw new MojoExecutionException("Selected version " +sdk +" is not available!");
@@ -289,7 +297,10 @@ public class XcodeBuild extends AbstractMojo implements PluginConstants {
 			init();
 			configure();
 			
-			executeAppCreateCommads();
+			if (!isCiBuild) {
+				executeAppCreateCommads();
+			}
+			
 			// below statement is used to get the project type. if it produces .a file, it static lib or app project
 			String projectTypeIdentified = getProjectType();
 			if (StringUtils.isEmpty(projectTypeIdentified)) {
@@ -560,14 +571,14 @@ public class XcodeBuild extends AbstractMojo implements PluginConstants {
 		}
 		
 		// if it is unit test, we have to set TEST_AFTER_BUILD=YES in build settings
-		// if (unittest) {
-		//	commands.add(TEST_AFTER_BUILD_YES);
-		// }
+		 if (unittest && !isXcodeLatestVersion) {
+			commands.add(TEST_AFTER_BUILD_YES);
+		 }
 		
 		// if it is unit test and application test, we need to pass this command
-		// if (unittest && applicationTest) {
-		// 	commands.add(RUN_UNIT_TEST_WITH_IOS_SIM_YES);
-		// }
+		 if (unittest && applicationTest && !isXcodeLatestVersion) {
+		 	commands.add(RUN_UNIT_TEST_WITH_IOS_SIM_YES);
+		 }
 		
 		if (StringUtils.isNotBlank(sdk) && !projectType.equals(MAC) && ((isXcodeLatestVersion && !isLogicalTest && !isApplicationTest) || !isXcodeLatestVersion)) {
 			commands.add(SDK);
@@ -650,10 +661,12 @@ public class XcodeBuild extends AbstractMojo implements PluginConstants {
 		try {
 			//if it is application test, no need to delete target directory
 			// To Delete the buildDirectory if already exists
-			if (buildDirectory.exists() && !applicationTest) {
-				getLog().info("removing exisiting target folder ... ");
-				FileUtils.deleteQuietly(buildDirectory);
-				buildDirectory.mkdirs();
+			if (!isCiBuild) {
+				if (buildDirectory.exists() && !applicationTest) {
+					getLog().info("removing exisiting target folder ... ");
+					FileUtils.deleteQuietly(buildDirectory);
+					buildDirectory.mkdirs();
+				}
 			}
 
 			buildInfoList = new ArrayList<BuildInfo>(); // initialization
